[CLI] move to entrypoint.
